requirements.txt
crewai==0.35.8
langchain-openai==0.1.8
python-dotenv==1.0.1 